General/requirements.txt
pandas==0.22.0
selenium==3.12.0
google-cloud==0.32.0

